backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psutil==6.0.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
cryptography==43.0.1
slowapi==0.1.9
websockets>=12.0
gunicorn==22.0.0
python-crontab==3.2.0
